P1_Structural_Profiling_Viz/requirements.txt
biopython
rich>=13.0.0
ViennaRNA>=2.5.0
